add support for circular deps + chore
