+++ gradle.properties
@@ -4,7 +4,7 @@ pluginGroup = works.szabope.pycharm
 pluginName = mypy
 pluginRepositoryUrl = https://github.com/szabope/mypy-pycharm-plugin
 # SemVer format -> https://semver.org
-pluginVersion = 1.0.10
+pluginVersion = 1.0.11
 
 # Supported build number ranges and IntelliJ Platform versions -> https://plugins.jetbrains.com/docs/intellij/build-number-ranges.html
 pluginSinceBuild = 243
